sql: consolidate shared SQL reader/status helpers and drop unused notice response binding
