+++ src/sql/postgres/protocol/StackReader.rs
@@ -1,101 +1,39 @@
 use crate::postgres::any_postgres_error::AnyPostgresError;
 use crate::postgres::protocol::new_reader::{NewReader, ReaderContext};
 use crate::shared::data::Data;
-use bun_core::strings;
+use crate::shared::stack_reader::{ShortRead, WrapReader};
 
-pub struct StackReader<'a> {
-    pub buffer: &'a [u8],
-    pub offset: &'a mut usize,
-    pub message_start: &'a mut usize,
+pub use crate::shared::stack_reader::StackReader;
+
+impl ShortRead for AnyPostgresError {
+    const SHORT_READ: Self = AnyPostgresError::ShortRead;
 }
 
-impl<'a> StackReader<'a> {
-    pub fn mark_message_start(&mut self) {
-        *self.message_start = *self.offset;
-    }
-
-    pub fn ensure_length(&self, length: usize) -> bool {
-        self.buffer.len() >= (*self.offset + length)
-    }
-
-    pub fn init(
-        buffer: &'a [u8],
-        offset: &'a mut usize,
-        message_start: &'a mut usize,
-    ) -> NewReader<StackReader<'a>> {
-        NewReader {
-            wrapped: StackReader {
-                buffer,
-                offset,
-                message_start,
-            },
-        }
-    }
-
-    pub fn peek(&self) -> &[u8] {
-        &self.buffer[*self.offset..]
-    }
-
-    pub fn skip(&mut self, count: usize) {
-        if *self.offset + count > self.buffer.len() {
-            *self.offset = self.buffer.len();
-            return;
-        }
-
-        *self.offset += count;
-    }
-
-    pub fn ensure_capacity(&self, count: usize) -> bool {
-        self.buffer.len() >= (*self.offset + count)
-    }
-
-    pub fn read(&mut self, count: usize) -> Result<Data, AnyPostgresError> {
-        let offset = *self.offset;
-        if !self.ensure_capacity(count) {
-            return Err(AnyPostgresError::ShortRead);
-        }
-
-        self.skip(count);
-        // Copy the &'a [u8] out before slicing so the returned Data borrows 'a,
-        // not &mut self.
-        let buffer: &'a [u8] = self.buffer;
-        Ok(Data::Temporary(bun_ptr::RawSlice::new(
-            &buffer[offset..*self.offset],
-        )))
-    }
-
-    pub fn read_z(&mut self) -> Result<Data, AnyPostgresError> {
-        // Inline `peek()` so `remaining` borrows 'a (via the Copy &'a [u8])
-        // instead of &self, allowing `self.skip()` below.
-        let buffer: &'a [u8] = self.buffer;
-        let remaining = &buffer[*self.offset..];
-        if let Some(zero) = strings::index_of_char(remaining, 0) {
-            let zero = zero as usize;
-            self.skip(zero + 1);
-            return Ok(Data::Temporary(bun_ptr::RawSlice::new(&remaining[0..zero])));
-        }
-
-        Err(AnyPostgresError::ShortRead)
+impl<'a> WrapReader<'a> for NewReader<StackReader<'a>> {
+    fn wrap(reader: StackReader<'a>) -> Self {
+        NewReader { wrapped: reader }
     }
 }
 
 impl<'a> ReaderContext for StackReader<'a> {
     fn mark_message_start(&mut self) {
-        Self::mark_message_start(self)
+        StackReader::mark_message_start(self)
     }
     fn peek(&self) -> &[u8] {
-        Self::peek(self)
+        StackReader::peek(self)
     }
     fn skip(&mut self, count: usize) {
-        Self::skip(self, count)
+        // The shared reader's signed skip clamps to the buffer end, matching
+        // the old unsigned behavior even when `count` exceeds `isize::MAX`.
+        StackReader::skip(self, isize::try_from(count).unwrap_or(isize::MAX))
     }
     fn ensure_length(&mut self, count: usize) -> bool {
-        Self::ensure_length(self, count)
+        StackReader::ensure_capacity(self, count)
     }
     fn read(&mut self, count: usize) -> Result<Data, AnyPostgresError> {
-        Self::read(self, count)
+        StackReader::read(self, count)
     }
     fn read_z(&mut self) -> Result<Data, AnyPostgresError> {
-        Self::read_z(self)
+        StackReader::read_z(self)
     }
 }
