+++ src/sql_jsc/lib.rs
@@ -20,6 +20,8 @@ pub mod shared {
     #[path = "CachedStructure.rs"]
     pub mod cached_structure;
 
+    pub mod connection_ctor_args;
+
     pub mod datetime_text;
 
     #[path = "ObjectIterator.rs"]
@@ -27,6 +29,8 @@ pub mod shared {
 
     #[path = "QueryBindingIterator.rs"]
     pub mod query_binding_iterator;
+
+    pub mod query_ctor_args;
 
     #[path = "SQLDataCell.rs"]
     pub mod sql_data_cell;
