+++ src/sql_jsc/postgres/protocol/error_response_jsc.rs
@@ -7,15 +7,11 @@ use bun_sql::postgres::protocol::field_message::FieldMessage;
 use crate::postgres::error_jsc::create_postgres_error;
 use bun_sql::postgres::any_postgres_error::PostgresErrorOptions;
 
-use super::notice_response_jsc::field_message_payload;
-
 pub(crate) fn to_js(this: &ErrorResponse, global_object: &JSGlobalObject) -> JSValue {
     let mut b = StringBuilder::default();
 
     for msg in this.messages.iter() {
-        // Every
-        // FieldMessage variant carries a single bun.String payload.
-        b.cap += field_message_payload(msg).utf8_byte_length() + 1;
+        b.cap += msg.payload().utf8_byte_length() + 1;
     }
     let _ = b.allocate();
